🚚 move register-selectors to selector-toolkits
