+++ packages/atom.io/internal/src/selector/index.ts
@@ -3,6 +3,6 @@ export * from "./create-standalone-selector"
 export * from "./create-writable-selector"
 export * from "./dispose-selector"
 export * from "./get-selector-dependency-keys"
-export * from "./register-selector"
+export * from "./selector-toolkits"
 export * from "./trace-selector-atoms"
 export * from "./update-selector-atoms"
